app/requirements.txt
aioredis==1.3.1
amqp==5.1.0
asgiref==3.5.0
astroid==2.9.3
async-generator==1.10
async-timeout==4.0.2
attrs==21.4.0
autobahn==22.3.2
Automat==20.2.0
autopep8==1.6.0
billiard==3.6.4.0
boto3==1.20.34
botocore==1.23.34
Brotli==1.0.9
cachetools==5.0.0
celery==5.2.3
certifi==2021.10.8
cffi==1.15.0
channels==3.0.4
channels-redis==3.3.1
charset-normalizer==2.0.12
click==8.1.0
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
ConfigArgParse==1.5.3
constantly==15.1.0
cryptography==36.0.2
daphne==3.0.2
defusedxml==0.7.1
Deprecated==1.2.13
diff-match-patch==20200713
dj-database-url==0.5.0
Django==4.0.2
django-cors-headers==3.11.0
django-debug-toolbar==3.2.4
django-filter==21.1
django-health-check==3.16.5
django-import-export==2.7.1
django-redis==5.2.0
django-storages==1.12.3
django-widget-tweaks==1.4.9
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
dnspython==2.2.1
docutils==0.18.1
drf-spectacular==0.21.1
email-validator==1.1.3
et-xmlfile==1.1.0
Faker==11.3.0
Flask==2.1.0
Flask-BasicAuth==0.2.0
Flask-Cors==3.0.10
flower==1.0.0
gevent==21.12.0
geventhttpclient==1.5.3
google-api-core==2.7.1
google-api-python-client==2.42.0
google-auth==2.6.2
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.1
googleapis-common-protos==1.56.0
greenlet==1.1.2
h11==0.13.0
hiredis==2.0.0
httplib2==0.20.4
humanize==4.0.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
inflection==0.5.1
iniconfig==1.1.1
isort==5.10.1
itsdangerous==2.1.2
Jinja2==3.1.1
jmespath==0.10.0
jsonschema==4.4.0
kombu==5.2.4
lazy-object-proxy==1.7.1
locust==2.5.1
Markdown==3.3.6
MarkupPy==1.14
MarkupSafe==2.0.1
mccabe==0.6.1
msgpack==1.0.3
oauthlib==3.2.0
odfpy==1.4.1
openpyxl==3.0.9
outcome==1.1.0
packaging==21.3
Pillow==9.0.1
platformdirs==2.5.1
pluggy==1.0.0
prometheus-client==0.13.1
prompt-toolkit==3.0.28
protobuf==3.19.4
psutil==5.9.0
psycopg2-binary==2.9.3
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
pyfcm==1.5.4
PyJWT==2.3.0
pylint==2.12.2
pyOpenSSL==22.0.0
pyotp==2.6.0
pyparsing==3.0.7
pyrsistent==0.18.1
pytest==7.1.1
pytest-django==4.5.2
python-dateutil==2.8.2
python-decouple==3.5
python-http-client==3.3.7
pytz==2022.1
PyYAML==6.0
pyzmq==22.3.0
redis==4.1.0
requests==2.27.1
requests-oauthlib==1.3.1
roundrobin==0.0.2
rsa==4.8
s3transfer==0.5.2
selenium==4.1.0
sendgrid==6.9.4
sentry-sdk==1.5.2
service-identity==21.1.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
sqlparse==0.4.2
starkbank-ecdsa==2.0.3
tablib==3.2.0
text-unidecode==1.3
toml==0.10.2
tomli==2.0.1
tornado==6.1
trio==0.20.0
trio-websocket==0.9.2
twilio==7.5.0
Twisted==22.2.0
txaio==22.2.1
typing_extensions==4.1.1
uritemplate==4.1.1
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
Werkzeug==2.1.0
whitenoise==5.3.0
wrapt==1.13.3
wsproto==1.1.0
xlrd==2.0.1
xlwt==1.3.0
zope.event==4.5.0
zope.interface==5.4.0
